Use dict.get to access card references
The references property is sometimes nonexistent on a card, and trying
to access it will throw a KeyError. I've replaced the card['references']
access with a call to dict.get, which defaults to an empty list. This
will prevent the endpoint from returning an error when it encounters a
card without and references.